Allow all unicode characters excepted "/" into unquoted names and value in requests.
This does not mean that a request like u"name=value" matches all items with
properties['name'] == 'value'. A string MUST be quoted to be regarded as a
string by kalamar.

Such a request will raise a RequestSyntaxError with message: "Failed to
convert value: u'value'.". Before, the message would be:
"Character u'$' is not allowed outside quotes."

This change make easier to add types to kalamar requests.