Fix issue #3 regarding FETCH_COUNT.
The FETCH_COUNT psql option makes use of the cursor API, which caches the
plan and the query.
The private state was not cacheable, since it was a custom data structure.
This patch encapsulates the private state in a Datum of type internal.
This should not pose any problem, since the content of the data structure itself
is NOT palloced (python objects).