Allow link flags required on OS X
By default, clang (which is what's actually invoked when you call gcc) will
insist that all symbols be resolved at link time, which is not what we want
here. The bug is actually in the link flags specified in PGXS; this is just
a workaround.

Tested on OS X 10.10.1 with clang-600.0.56. Fixes #88.