Take various databases support for NULLS ordering
Before that, NULLS FIRST, NULLS LAST was systematically appended to every
order by clause.

Now, The sqlalchemy FDW tries to avoid generating such a clause if it is not
necessary. In addition to that, if the remoted database does not support
the clause, AND its default behavior differs from the one the query is expected
to perform, the sort is not pushed down at all.

Thanks to Kyle Lilly for the report !

Fixes #126